webui/aws_infra/pyproject.toml
[project]
name = "webui-aws_infra"
version = "0.1.0"
description = "PAVI webUI AWS resource definitions"
readme = "README.md"
authors = [
    {name = "Manuel Luypaert", email = "luypaert@ebi.ac.uk"}
]

requires-python = "==3.12.*"
dependencies = [
    # CDK requirements
    "aws-cdk-lib==2.*",
    # CDK Alpha constructs for Amplify
    "aws-cdk.aws-amplify-alpha==2.*",
    # PAVI shared AWS module
    "pavi_shared_aws @ file:///tmp/pavi_shared_aws-0.0.0-py3-none-any.whl"
]

[project.optional-dependencies]
test = [
    "flake8==7.3.*",
    "flake8-unused-arguments==0.0.*",
    "mypy==1.18.*",
    "pytest==8.4.*"
]

[tool.setuptools.packages.find]
exclude = ["node_modules/*"]
